pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
authors = [
    {name = "sengh",email = "jaswindersingh11061998@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
google-adk = "^1.12.0"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
